refactor: split MCP server construction
